+++ language/script/src/main/kotlin/org/pcsoft/intellij/plugin/inno_setup/script/language/file_type/lang/IsLanguageFileIconProvider.kt
@@ -17,7 +17,7 @@ import com.intellij.openapi.project.Project
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.psi.PsiManager
 import org.pcsoft.intellij.plugin.inno_setup.script.language.feature.IsLanguageIconHelper
-import org.pcsoft.intellij.plugin.inno_setup.script.language.file_type.IsIcons
+import org.pcsoft.intellij.plugin.inno_setup.script.IsIcons
 import org.pcsoft.intellij.plugin.inno_setup.script.language.file_type.IsScriptFile
 import javax.swing.Icon