+++ plugin/src/main/kotlin/org/pcsoft/intellij/plugin/inno_setup/language/feature/completion/IsSectionCompletionContributor.kt
@@ -108,5 +108,3 @@ class IsSectionCompletionContributor : CompletionContributor() {
         )
     }
 }
-
-