+++ plugin/src/main/kotlin/org/pcsoft/intellij/plugin/inno_setup/language/feature/completion/provider/IsSectionEnvVarAfterPercentProvider.kt
@@ -18,7 +18,7 @@ import com.intellij.codeInsight.completion.CompletionResultSet
 import com.intellij.codeInsight.completion.PrioritizedLookupElement
 import com.intellij.codeInsight.lookup.LookupElementBuilder
 import com.intellij.util.ProcessingContext
-import org.pcsoft.intellij.plugin.inno_setup.script.language.file_type.IsIcons
+import org.pcsoft.intellij.plugin.inno_setup.script.IsIcons
 
 /**
  * Completion of environment variable names inside `{%…}` constants. Looks back from the caret